merge

+++ include/vector.h
@@ -1,9 +1,8 @@
 #pragma once
 #include <stdint.h>
-//#include "arm_math.h"
 
 typedef struct coordinate{
-    int8_t x, y;
+    int32_t x, y;
 } coordinate;
 
 coordinate sumVector(coordinate vector_a, coordinate vector_b);
